move TraverseFlow to context.rs
